Move trace-to-DataFrame functionality to module
