+++ pymc3/__init__.py
@@ -13,6 +13,7 @@ from .tuning import *
 from .debug import *
 
 from .diagnostics import *
+from .backends.tracetab import *
 
 from .plots import *
