+++ pymc3/diagnostics.py
@@ -4,7 +4,7 @@ import numpy as np
 from .stats import autocorr, autocov, statfunc
 from copy import copy
 
-__all__ = ['geweke', 'gelman_rubin', 'trace_to_dataframe', 'effective_n']
+__all__ = ['geweke', 'gelman_rubin', 'effective_n']
 
 
 @statfunc
@@ -264,12 +264,3 @@ def effective_n(mtrace):
             n_eff[var] = [calc_n_eff(y.transpose()) for y in x.transpose()]
 
     return n_eff
-    
-
-def trace_to_dataframe(trace):
-    """Convert a PyMC trace consisting of 1-D variables to a pandas DataFrame
-    """
-    import pandas as pd
-    return pd.DataFrame(
-        {varname: np.squeeze(trace.get_values(varname, combine=True))
-         for varname in trace.varnames})
